Split logic into separate files
