+++ crates/driver/src/infra/mod.rs
@@ -11,7 +11,7 @@ pub mod simulator;
 pub mod solver;
 pub mod time;
 pub mod tokens;
-
+/// TODO put bad tokens here?
 pub use {
     self::solver::Solver,
     api::Api,
