+++ crates/driver/src/infra/solver/mod.rs
@@ -124,8 +124,7 @@ pub struct Config {
     /// Which `tx.origin` is required to make quote verification pass.
     pub quote_tx_origin: Option<eth::Address>,
     pub response_size_limit_max_bytes: usize,
-    /// Tokens that are explicitly allow- or deny-listed.
-    pub tokens_supported: HashMap<eth::TokenAddress, bad_tokens::Quality>,
+    pub bad_token_detection: BadTokenDetection,
 }
 
 impl Solver {
@@ -154,8 +153,8 @@ impl Solver {
         })
     }
 
-    pub fn tokens_supported(&self) -> &HashMap<eth::TokenAddress, bad_tokens::Quality> {
-        &self.config.tokens_supported
+    pub fn bad_token_detection(&self) -> &BadTokenDetection {
+        &self.config.bad_token_detection
     }
 
     pub fn persistence(&self) -> Persistence {
@@ -301,4 +300,11 @@ impl Error {
             _ => false,
         }
     }
+}
+
+#[derive(Debug, Clone)]
+pub struct BadTokenDetection {
+    /// Tokens that are explicitly allow- or deny-listed.
+    pub tokens_supported: HashMap<eth::TokenAddress, bad_tokens::Quality>,
+    pub enable_simulation_based_bad_token_detection: bool,
 }
