+++ crates/driver/src/run.rs
@@ -1,6 +1,6 @@
 use {
     crate::{
-        domain::Mempools,
+        domain::{competition::bad_tokens, Mempools},
         infra::{
             self,
             blockchain::{self, Ethereum},
@@ -66,10 +66,13 @@ async fn run_with(args: cli::Args, addr_sender: Option<oneshot::Sender<SocketAdd
             eth.clone(),
         )
         .unwrap(),
+        bad_token_detector: bad_tokens::simulation::Detector::new(
+            config.simulation_bad_token_max_age,
+            &eth,
+        ),
         eth,
         addr: args.addr,
         addr_sender,
-        bad_token_detection: config.bad_token_detection,
     }
     .serve(
         async {
